Protein denaturation

PDBTools.CreamerDenaturedModel — Type

CreamerDenaturedModel

Type that specifies that a m-value calculation will consider a Creamer model of protein denaturation. This type is used as the first input variable of the mvalue function.

Construction:

CreamerDenaturedModel(atoms::AbstractVector{<:Atom})

and, optionally, the second argument is the type of denatured model to be used, for example:

CreamerDenaturedModel(atoms::AbstractVector{<:Atom}, 1)

where the type parameter can be 1, 2, or 3, for the minimal, average, and maximal denatured estimated accessible surface ares of Creamer.

Use the CreamerDenaturedModel model as the first input argument of mvalue, for example:

mvalue(CreamerDenaturedModel(prot), "urea")

to obtain the estimated m-value of denaturation in "urea".

Reference:

Creamer TP, Srinivasan R, Rose GD. Modeling unfolded states of proteins and peptides. II. Backbone solvent accessibility. Biochemistry. 1997;36:2832–2835. doi: 10.1021/bi962819o.

The optional sasa_parameterization keyword defines which denatured SASA parameterization will be used, with the published Creamer SASAs (:original - default) or the recomputed parameters based on the CATH S20 classification (:cath_s20).

PDBTools.mvalue — Method

mvalue(m::CreamerDenaturedModel, cosolvent::AbstractString; model=AutonBolen)

Computes an estimate of the effect of a cosolvent on the m-value of denaturation of a protein.

Example

julia> using PDBTools

julia> prot = read_pdb(PDBTools.TESTPDB, "protein");

julia> m = mvalue(CreamerDenaturedModel(prot), "urea");

julia> m.tot
-1.290518033485419

The estimate of the m-value associated to protein denaturation (the change in transfer free energy when a protein undergoes denaturation) can be (and is usually) computed using the estimates of exposed residue surface area obtained by Creamer.

Here, we compute the m-value estimates using the Creamer model, but first wrapping the atom array in the CreamerDenaturedModel type, which defines also the denaturation extent to be considered, by default, "mean":

using PDBTools
prot = read_pdb(PDBTools.TESTPDB, "protein")
model = CreamerDenaturedModel(prot)

CreamerDenaturedModel of a 1463-atom protein and mean denaturation.

A second argument of the constructor defines the extent to the "minimal", "mean", "maximal", as described in the docstring below.

The m-value of denaturation in a cosolvent can be computed, then, with:

m = mvalue(model, "urea"; model=AutonBolen)
println("m-value of denaturation = ", m.tot)

m-value of denaturation = -1.290518033485419

where the output is a dictionary containing the total, backbone, side-chain, and residue-type specific contributions to the transfer free energies, in kcal/mol.