Protein transfer free energy and m-values

These function compute the transfer free energies of proteins from water to different solvent, using the Tanford additive transfer models.

m-values are the transfer free-energy difference, here in kcal/mol, between two structures from water to a 1M solution of a cosolvent. Estimates of m-values for denaturation events can be computed using the Creamer estimates for denatued accessible surface areas.

PDBTools.transfer_free_energy — Function

transfer_free_energy(atoms::AbstractVector{<:PDBTools.Atom}, cosolvent::AbstractString; kargs...)
transfer_free_energy(sasa_atoms::SASA{3,<:AbstractVector{<:Atom}}, cosolvent::AbstractString; kargs...)

Calculates the transfer free energy (in 1M solution, in kcal/mol) using the Tanford transfer model, as implemented by Moeser and Horinek [1] or by Auton and Bolen [2,3].

Positional Arguments

atoms:: Atoms of the system (a vector of PDBTools.Atom objects)

or

sasa_atoms::SASA{3,<:AbstractVector{<:Atom}}: SASA object with solvent accessible surface area of the atoms.

and

cosolvent::AbstractString: The cosolvent to consider.

Available models and cosolvents for each model:

MoeserHorinek: "urea"

AutonBolen: "betaine", "proline", "sarcosine", "sorbitol", "sucrose", "tmao", "urea", "urea-app"

MoeserHorinekFit: "betaine", "proline", "sarcosine", "sorbitol", "sucrose", "tmao", "urea"

Keyword Arguments (optional)

model::Type{<:MValueModel}=AutonBolen: The model to use for the calculation.

sel::Union{String,Function}=all: Selection of atoms to consider in the calculation. Can be a selection string or a function that takes an Atom and returns a Bool.

backbone::Function = PDBTools.isbackbone: Function to identify backbone atoms.

sidechain::Function = PDBTools.issidechain: Function to identify side chain atoms.

parallel:Bool = true: Set parallelization, requires starting Julia multithreaded.

Returns

A TransferFreeEnergy object, with fields:

ntatoms::Int: Number of atoms considered.

tot::Float32: Total m-value (kcal/mol/M).

bb::Float32: Backbone contribution to the m-value (kcal/mol/M).

sc::Float32: Side chain contribution to the m-value (kcal/mol/M).

residue_contributions_bb::Vector{Float32}: Backbone contributions of each residue to the m-value.

residue_contributions_sc::Vector{Float32}: Side-chain contributions of each residue to the m-value.

cosolvent::String: The cosolvent considered.

Example

using PDBTools
# directly from the atoms:
prot = read_pdb("native.pdb")
transfer_free_energy(prot, "urea")
# or providing the SASA object:
sasa_prot = sasa_particles(prot)
transfer_free_energy(sasa_prot, "urea")

Warning

When providing the vector of atoms, the SASA calculation will be performed using Creamer united atom types. On the other hand, if the SASAs are computed previously, they will correspond to the method used which, by default, consider all atoms of the structure, including hydrogens, and uses standard vdW atoms.

References

https://doi.org/10.1021/jp409934q

https://doi.org/10.1016/s0076-6879(07)28023-1

https://www.pnas.org/doi/10.1073/pnas.0706251104

PDBTools.mvalue — Function

mvalue(
    sasa_initial::SASA{3,<:AbstractVector{<:Atom}},    
    sasa_final::SASA{3,<:AbstractVector{<:Atom}},
    cosolvent::AbstractString;
    sel::Union{String,Function}=all,
    model::Type{<:MValueModel}=AutonBolen,
    backbone::Function = isbackbone,
    sidechain::Function = issidechain,
    parallel:Bool = true,
)

Calculates the m-value (transfer free energy of a protein in 1M solution, in kcal/mol) using the Tanford transfer model, as implemented by Moeser and Horinek [1] or by Auton and Bolen [2,3].

Positional Arguments

sasa_initial::SASA{3,<:AbstractVector{<:Atom}}: SASA object representing the initial state (e.g., native state).

sasa_final::SASA{3,<:AbstractVector{<:Atom}}: SASA object representing the final state (e.g., denatured state).

cosolvent::AbstractString: The cosolvent to consider (case insensitive).

Available models and cosolvents for each model:

MoeserHorinek: "urea"

AutonBolen: "betaine", "proline", "sarcosine", "sorbitol", "sucrose", "tmao", "urea", "urea-app"

MoeserHorinekFit: "betaine", "proline", "sarcosine", "sorbitol", "sucrose", "tmao", "urea"

Keyword Arguments

sel::Union{String,Function}=all: Selection of atoms to consider in the calculation. Can be a selection string or a function that takes an Atom and returns a Bool.

model::Type{<:MValueModel}=AutonBolen.

backbone::Function = PDBTools.isbackbone: Function to identify backbone atoms.

sidechain::Function = PDBTools.issidechain: Function to identify side chain atoms.

parallel:Bool = true: Set parallelization, requires starting Julia multithreaded.

Returns

A MValue object, with fields:

ntatoms::Int: Number of atoms considered.

tot::Float32: Total m-value (kcal/mol/M).

bb::Float32: Backbone contribution to the m-value (kcal/mol/M).

sc::Float32: Side chain contribution to the m-value (kcal/mol/M).

residue_contributions_bb::Vector{Float32}: Backbone contributions of each residue to the m-value.

residue_contributions_sc::Vector{Float32}: Side-chain contributions of each residue to the m-value.

cosolvent::AbstractString: The cosolvent considered.

Example

using PDBTools

initial_state = read_pdb("native.pdb")
final_state = read_pdb("desnat.pdb")

sasa_initial = sasa_particles(initial_state)
sasa_final = sasa_particles(final_state)

mvalue(sasa_initial, sasa_final, "chain A"; model=AutonBolen, cosolvent="TMAO")

References

https://doi.org/10.1021/jp409934q

https://doi.org/10.1016/s0076-6879(07)28023-1

https://www.pnas.org/doi/10.1073/pnas.0706251104

mvalue(m::CreamerDenaturedModel, cosolvent::AbstractString; model=AutonBolen)

Computes an estimate of the effect of a cosolvent on the m-value of denaturation of a protein.

Example

julia> using PDBTools

julia> prot = read_pdb(PDBTools.TESTPDB, "protein");

julia> m = mvalue(CreamerDenaturedModel(prot), "urea");

julia> m.tot
-1.290518033485419

Here we implement the Moeser/Horinek, Auton/Bolen, and a version of Moeser/Horinek with Gly-non-ideality corrections (see this section) models. (1, 2, 3). Typically, these models are used to obtain the effect of cosolvent on the structural stability of proteins, but the current implementation allows the practical use of these functions to compute m-values of more general transformations, as described in the examples.

The transfer free energy of a protein from water to a 1M solution of a cosolvent can be estimated with the transfer_free_energy function:

using PDBTools
native_state = read_pdb(PDBTools.src_dir*"/tools/mvalue/1MJC_native.pdb", "protein")
tfe = transfer_free_energy(native_state, "urea")

PDBTools.TransferFreeEnergy{AutonBolen} - 69 residues to 1M "urea".
    Total transfer free energy: -0.7395303 kcal mol⁻¹
    Backbone contributions: -0.98496294 kcal mol⁻¹
    Side-chain contributions: 0.2454326 kcal mol⁻¹

The resulting TransferFreeEnegy object contains the information of the contribution of each residue to the transfer free energy obtained, split into backbone and side-chain contributions:

tfe.residue_contributions_bb[1]

-0.086069845f0

tfe.residue_contributions_sc[1]

0.011580416f0

When multiple protein conformations are of interest, the mvalue methods provide a direct way to compute the variations in transfer free energies associated to the states involved, as shown in the following examples.